Fixing path for JS assets on Mailchimp form
The javascript files used for the Mailchimp signup form are not located
in the bundles directory. This commit fixes the paths being generated
for the template context in which they are loaded.

Addresses #472

Move custom.js content into public.js

Lint
