+++ cms/templates/cms/home_page.html
@@ -264,7 +264,4 @@
   </div>
 </div>
 <script type="text/javascript" src="{{ public_src }}"></script>
-<script type="text/javascript" src="{{ wow_src }}"></script>
-<script type="text/javascript" src="{{ ajaxchimp_src }}"></script>
-<script type="text/javascript" src="{{ custom_src }}"></script>
 {% endblock %}
