+++ package.json
@@ -2,6 +2,7 @@
   "name": "micromasters",
   "version": "0.0.0",
   "dependencies": {
+    "ajaxchimp": "^1.3.0",
     "babel": "6.5.2",
     "babel-core": "6.8.0",
     "babel-eslint": "6.0.4",
@@ -21,6 +22,7 @@
     "fetch-mock": "4.4.0",
     "file-loader": "0.8.5",
     "history": "2.1.1",
+    "imports-loader": "^0.6.5",
     "iso-3166-2": "0.4.0",
     "isomorphic-fetch": "2.2.1",
     "istanbul": "1.0.0-alpha.2",
@@ -56,7 +58,8 @@
     "url-loader": "0.5.7",
     "warning": "2.1.0",
     "webpack": "1.13.0",
-    "webpack-dev-server": "1.14.1"
+    "webpack-dev-server": "1.14.1",
+    "wowjs": "^1.1.3"
   },
   "scripts": {
     "postinstall": "./webpack_if_prod.sh",
